fix: update discussion API permissions
